requirements.txt
djangorestframework==3.14.0
environs==10.3.0
pip-chill==1.0.3
psycopg2==2.9.9
Django==5.0.2
